ui/requirements.txt
streamlit==1.32.2
requests==2.31.0
streamlit-pdf-viewer==0.0.28
pdf2image==1.17.0
streamlit-image-coordinates==0.1.6
Pillow==10.2.0